Add files via upload
My Web Final For GIT-215 
I still need to add some content but major fixes right now...
Need to make the two half width cards on the home page display next to each other.
Need to make the dark blue wrapper fill the edges on the homepage. 
Need to fix the whole site width to 980px or I can choose to make the site responsive but it has to be fully responsive and not half-assed.